feat: Home화면 API Model 추가
