+++ Todorama/Network/Model/Popular.swift
@@ -1,0 +1,16 @@
+//
+//  Popular.swift
+//  Todorama
+//
+//  Created by 최정안 on 3/21/25.
+//
+
+import Foundation
+struct Popular: Decodable, Hashable {
+    let results: [PopularDetail]
+}
+struct PopularDetail: Decodable, Hashable {
+    let id: Int
+    let poster: String
+}
+
